requirements.txt
gradio>=3.0
python-dotenv
numpy
cohere
weaviate-client
groq
google-genai
openai